pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.116.1",
    "sqlalchemy==2.0.42",
    "psycopg2==2.9.10",
    "pydantic==2.11.7",
    "uvicorn==0.35.0",
    "xlsxwriter==3.2.5",
    "python-dotenv==1.1.1"
]

[tool.uv.workspace]
members = [
    "git",
]
